+++ include/ccf/crypto/verifier.h
@@ -185,11 +185,13 @@ namespace crypto
      * @param trusted_certs Vector of trusted certificates
      * @param chain Vector of ordered untrusted certificates used to
      *  build a chain to trusted certificates
+     * @param ignore_time Flag to disable certificate expiry checks
      * @return true if the verification is successfull
      */
     virtual bool verify_certificate(
       const std::vector<const Pem*>& trusted_certs,
-      const std::vector<const Pem*>& chain = {}) = 0;
+      const std::vector<const Pem*>& chain = {},
+      bool ignore_time = false) = 0;
 
     /** Indicates whether the certificate (held intenally) is self-signed */
     virtual bool is_self_signed() const = 0;