+++ src/host/main.cpp
@@ -424,6 +424,16 @@ int main(int argc, char** argv)
       LOG_INFO_FMT("Creating new node - recover");
       startup_config.initial_service_certificate_validity_days =
         config.command.recover.initial_service_certificate_validity_days;
+      auto idf = config.command.recover.previous_service_identity_file;
+      if (!files::exists(idf))
+      {
+        throw std::logic_error(fmt::format(
+          "Recovery requires a previous service identity certificate; cannot "
+          "open '{}'",
+          idf));
+      }
+      LOG_INFO_FMT("Reading previous service identity from {}", idf);
+      startup_config.recover.previous_service_identity = files::slurp(idf);
     }
     else
     {