+++ src/node/node_state.h
@@ -230,7 +230,8 @@ namespace ccf
         hooks,
         &view_history,
         true,
-        config.startup_snapshot_evidence_seqno_for_1_x);
+        config.startup_snapshot_evidence_seqno_for_1_x,
+        config.recover.previous_service_identity);
 
       startup_seqno = startup_snapshot_info->seqno;
       ledger_idx = startup_seqno.value();
@@ -415,6 +416,13 @@ namespace ccf
         }
         case StartType::Recover:
         {
+          if (!config.recover.previous_service_identity)
+          {
+            throw std::logic_error(
+              "Recovery requires the certificate of the previous service "
+              "identity");
+          }
+
           network.identity = std::make_unique<ReplicatedNetworkIdentity>(
             curve_id,
             config.startup_host_time,
@@ -587,7 +595,8 @@ namespace ccf
                 hooks,
                 &view_history,
                 resp.network_info->public_only,
-                startup_snapshot_info->evidence_seqno);
+                startup_snapshot_info->evidence_seqno,
+                config.recover.previous_service_identity);
 
               for (auto& hook : hooks)
               {
@@ -1236,7 +1245,8 @@ namespace ccf
           hooks,
           &view_history,
           false,
-          startup_snapshot_info->evidence_seqno);
+          startup_snapshot_info->evidence_seqno,
+          config.recover.previous_service_identity);
         startup_snapshot_info.reset();
       }
 
@@ -1282,7 +1292,9 @@ namespace ccf
       RINGBUFFER_WRITE_MESSAGE(AppMessage::launch_host_process, to_host, json);
     }
 
-    void transition_service_to_open(kv::Tx& tx) override
+    void transition_service_to_open(
+      kv::Tx& tx,
+      AbstractGovernanceEffects::ServiceIdentities identities) override
     {
       std::lock_guard<std::mutex> guard(lock);
 
@@ -1303,6 +1315,31 @@ namespace ccf
         LOG_DEBUG_FMT(
           "Service in state {} is already open", service_info->status);
         return;
+      }
+
+      if (
+        service_info->status == ServiceStatus::RECOVERING &&
+        (!config.recover.previous_service_identity ||
+         !identities.previous.has_value()))
+      {
+        throw std::logic_error(
+          "Recovery with service certificates requires both, a previous "
+          "service identity certificate during node startup and a "
+          "transition_service_to_open proposal that contains previous and next "
+          "service certificates");
+      }
+
+      if (identities.next != service_info->cert)
+      {
+        throw std::logic_error(
+          "Service identity mismatch: the next service identity in the "
+          "transition_service_to_open proposal does not match the current "
+          "service identity");
+      }
+
+      if (identities.previous)
+      {
+        service_info->previous_service_identity = *identities.previous;
       }
 
       if (is_part_of_public_network())